Reimplement Causeway functionality

+++ rest/src/main/java/org/jboss/pnc/causeway/rest/BrewBuild.java
@@ -5,19 +5,19 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class BrewBuild extends BrewNVR {
 
-    private final Long id;
+    private final Integer id;
 
     @JsonCreator
-    public BrewBuild(@JsonProperty("id")Long id, @JsonProperty("name")String name, @JsonProperty("version")String version, @JsonProperty("release")String release) {
+    public BrewBuild(@JsonProperty("id")Integer id, @JsonProperty("name")String name, @JsonProperty("version")String version, @JsonProperty("release")String release) {
         this(id, new BrewNVR(name, version, release));
     }
 
-    public BrewBuild(Long id, BrewNVR nvr) {
+    public BrewBuild(Integer id, BrewNVR nvr) {
         super(nvr.getName(), nvr.getRelease(), nvr.getVersion());
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
